docs: add Q,R,S & T problem

+++ H_Dots.py
@@ -1,4 +1,4 @@
 import math
-n=[int(i) for i in input().split()]
+n=[float(i) for i in input().split()]
 r=math.sqrt((n[0]-n[2])**2+(n[1]-n[3])**2)
 print("{0:.3f}".format(r))
